requirements.txt
SpiNNUtilities >= 1!4.0.0, < 1!5.0.0
SpiNNMachine >= 1!4.0.0, < 1!5.0.0
SpiNNMan >= 1!4.0.0, < 1!5.0.0
SpiNNStorageHandlers >= 1!4.0.0, < 1!5.0.0
SpiNNaker_PACMAN >= 1!4.0.0, < 1!5.0.0
SpiNNaker_DataSpecification >= 1!4.0.0, < 1!5.0.0
spalloc >= 1.0.0, < 2.0.0
SpiNNFrontEndCommon >= 1!4.0.0, < 1!5.0.0
sPyNNaker >= 1!4.0.0, < 1!5.0.0
SpyNNaker8 >= 1!4.0.0, < 1!5.0.0
matplotlib==2.2.2
networkx==2.1
prettytable==0.7.2
